// frontend/lib/api/model/actionResponseSchema.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { ActionType } from './actionType';

export interface ActionResponseSchema {
  /**
   * @minimum -2147483648
   * @maximum 2147483647
   */
  id: number;
  /**
   * EMAIL: email
   * @maxLength 20
   */
  type: ActionType;
  /** @maxLength 255 */
  target: string;
  /** @maxLength 255 */
  value: string;
  /**
   * @minimum -2147483648
   * @maximum 2147483647
   */
  automation_id: number;
}